backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.8.2
psycopg[binary]==3.2.1
SQLAlchemy==2.0.32
pgvector==0.3.3
PyMuPDF==1.24.9
Pillow==10.4.0
openai>=1.30.0
